Move Override tests to separate file
